requirements.txt
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
matplotlib==3.8.3
apache-airflow==2.8.3
